# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Users",
    "version": "1.0",
    "description": "Test project",
    "contact": {
      "name": "Vika",
      "url": "https://github.com/HladkaViktoriia",
      "email": "vktr.gldk10.98@gmail.com"
    }
  },
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "getAllUsers",
        "responses": {
          "200": {
            "description": "Get all users",
            "schema": {
              "$ref": "#/definitions/singleUser"
            }
          },
          "400": {
            "description": "Just an error",
            "schema": {
              "$ref": "#/definitions/status_400"
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "createUser",
        "description": "This is how to create all users",
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "default": "Vika",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "default": "vika@gmail.com",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "default": "123456",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User was created",
            "schema": {
              "$ref": "#/definitions/singleUser"
            }
          },
          "400": {
            "description": "Just an error",
            "schema": {
              "$ref": "#/definitions/status_400"
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "user"
        ],
        "operationId": "getUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Get a user by id",
            "schema": {
              "$ref": "#/definitions/singleUser"
            }
          },
          "400": {
            "description": "Just an error",
            "schema": {
              "$ref": "#/definitions/status_400"
            }
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "operationId": "updateUser",
        "description": "This is how to update a user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "required": true
          },
          {
            "name": "Authorization",
            "in": "header",
            "default": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "default": "Vika",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User was updated",
            "schema": {
              "$ref": "#/definitions/singleUser"
            }
          },
          "400": {
            "description": "Just an error",
            "schema": {
              "$ref": "#/definitions/status_400"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "operationId": "deleteUser",
        "description": "This is how to delete a user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "required": true
          },
          {
            "name": "Authorization",
            "in": "header",
            "default": "",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Get a user by id",
            "schema": {
              "$ref": "#/definitions/singleUser"
            }
          },
          "400": {
            "description": "Just an error",
            "schema": {
              "$ref": "#/definitions/status_400"
            }
          }
        }
      }
    },
    "/auth": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "authUser",
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "default": "vika@gmail.com",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "default": "123456",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Get all users",
            "schema": {
              "$ref": "#/definitions/singleUser"
            }
          },
          "400": {
            "description": "Just an error",
            "schema": {
              "$ref": "#/definitions/status_400"
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "logout"
        ],
        "operationId": "logoutUser",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "default": "",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Get all users",
            "schema": {
              "$ref": "#/definitions/singleUser"
            }
          },
          "400": {
            "description": "Just an error",
            "schema": {
              "$ref": "#/definitions/status_400"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "singleUser": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "123456test",
          "description": "MongoDB ID"
        },
        "email": {
          "type": "string",
          "example": "vika@gmail.com"
        },
        "name": {
          "type": "string",
          "example": "Vika"
        }
      }
    },
    "status_400": {
      "type": "object",
      "properties": {
        "msg": {
          "type": "string",
          "example": "Something went wrong"
        },
        "code": {
          "type": "number",
          "example": 400
        }
      }
    }
  }
}